Move requirements to Operator
